Handle differences in Crypt::JWT from JSON::WebToken
Pass allow_none to encode_jwt() and decode_jwt().
(Crypt::JWT doesn't allow alg 'none' by default.  Could make it an option
of OIDC::Lite, but this is fine for now.)

For RSA algorithms, if key is a scalar, pass it as a reference to the
scalar.
(Crypt::JWT requires RSA keys to be a reference, because of how it uses
Crypt::PK::RSA.  It would be better if this conversion were done in
Crypt::JWT instead.)

Make test of an invalid signature with alg 'none' a TODO test.
(Crypt::JWT allows any signature if alg is 'none'.  JSON::WebToken requires
signature to be blank if alg is 'none'.)